LPD-77438 Extract logic from component
